Direct request PricePackage

+++ aspnet-core/src/TACHYON.Application.Shared/PricePackages/Dto/NormalPricePackage/CreateOrEditNormalPricePackageDto.cs
@@ -13,12 +13,12 @@ namespace TACHYON.PricePackages.Dto.NormalPricePackage
         public int TransportTypeId { get; set; }
         public long TrucksTypeId { get; set; }
         [Range(PricePackagesConst.MinPriceNumber, PricePackagesConst.MaxPriceNumber)]
-        public float DirectRequestPrice { get; set; }
+        public decimal DirectRequestPrice { get; set; }
         [Range(PricePackagesConst.MinPriceNumber, PricePackagesConst.MaxPriceNumber)]
-        public float MarcketPlaceRequestPrice { get; set; }
+        public decimal MarcketPlaceRequestPrice { get; set; }
         [Range(PricePackagesConst.MinPriceNumber, PricePackagesConst.MaxPriceNumber)]
-        public float TachyonMSRequestPrice { get; set; }
-        public float? PricePerExtraDrop { get; set; }
+        public decimal TachyonMSRequestPrice { get; set; }
+        public decimal? PricePerExtraDrop { get; set; }
         public bool IsMultiDrop { get; set; }
         [Required]
         public int OriginCityId { get; set; }
